Extract out body parser spec
